requirements.txt
matplotlib==3.8.0
numpy==1.21.6
Pillow==10.4.0
scipy==1.14.1
torch==2.4.1
torchvision==0.17.2
tqdm==4.66.1
tensorboard==2.17.1